Documentation

Types

Missing docstring.

Missing docstring for ParabBand(mstar,ϵ₀,type,deg). Check Documenter's build log for details.

Missing docstring.

Missing docstring for BandStructure(n,bands,μ). Check Documenter's build log for details.

Functions

Missing docstring.

Missing docstring for electrical_conductivity(bandstructure::BandStructure, temperature, scattering_time; exportasdf::Bool, fulltensor::Bool). Check Documenter's build log for details.

Missing docstring.

Missing docstring for seebeck_coefficient(bandstructure::BandStructure, temperature, scattering_time; exportasdf::Bool, fulltensor::Bool). Check Documenter's build log for details.

Missing docstring.

Missing docstring for thermal_conductivity(bandstructure::BandStructure, temperature, scattering_time; exportasdf::Bool, fulltensor::Bool). Check Documenter's build log for details.

Missing docstring.

Missing docstring for carrier_concentration(bandstructure::BandStructure, temperature, fermi_level; exportasdf::Bool). Check Documenter's build log for details.

Missing docstring.

Missing docstring for lorenz_tensor(bandstructure::BandStructure, temperature, scattering_time; exportasdf::Bool, fulltensor::Bool). Check Documenter's build log for details.

Missing docstring.

Missing docstring for Ln(ϵ₀, μ, bandtype, T, β, n, idx, τ). Check Documenter's build log for details.

Missing docstring.

Missing docstring for constant(A=1.0). Check Documenter's build log for details.

Mstar2t.Scattering.T_fun — Method

T_fun(τ::Function)

Set the temperature dependence of the relaxation time from a function given by the user.

Example: f(T) = sqrt(T) # τ ∝ √T τmodel = Tfun(f)

Mstar2t.Scattering.impurity — Function

impurity(ϵ_im::Real, A_im::Real=1;γ::Real=1.)

Include impurity scattering in the simulations.

Parameters: ϵim: energy of the impurity in eV A: multiplicative constant in front of the functional expression γ: γ-parameter three-parameter Lorentzian function (Ref: https://en.wikipedia.org/wiki/Cauchydistribution)

Example: ϵim = 0.2 # eV τmodel = impurity(ϵ_im,A=1e-1)

Missing docstring.

Missing docstring for acoustic(bands_min, A_sm, τm_max; T₀, μ_min, μ_max). Check Documenter's build log for details.

Mstar2t.Scattering.matthiessen — Method

matthiessen(τ_models::Array{ScModel};γ::Float64=-1.)

This function applies Matthiessen's rule to sum up different scattering mechanisms.

Parameters: τmodels: vector of relaxation time models γ: exponent in the generalized Matthiessen's rule: τ = (\sumi τi^γ)^(1/γ)$, where each \taui$ can be a function of $T$ and/or $μ$.

Example: ϵim = 0.2 # eV τ1 = constant() τ2 = impurity(ϵim,A=1e-1) τ_model = matthiessen([τ1,τ2])

Missing docstring.

Missing docstring for plot(num_plots, x, y, z=Float64[]; titles, xlabels, ylabels, zlabel, color, colorscheme, vlines, annotations). Check Documenter's build log for details.

@docs plot_bandstructure(bs,xaxis=range(-1,1,length=100); colors=nothing)

@docs plot_τ(scm, t, type="e-T"; μ, ϵ₀, bandtype)

@docs savedata(path, data)

@docs savefig(fullpath, fig)


## Index

@index ```